quiz app challenge refactoring the project files
